Break SearchInstance into its own file
